+++ src/provider/providerWrapper.ts
@@ -61,22 +61,32 @@ export class ProviderWrapper {
   }
 
   async getErc725YVersion(address: string): Promise<ERC725_VERSION> {
-    const isErc725Y = await this.supportsInterface(
+    const isErc725Yv5 = await this.supportsInterface(
+      address,
+      ERC725Y_INTERFACE_IDS['5.0'],
+    );
+
+    if (isErc725Yv5) {
+      return ERC725_VERSION.ERC725_v5;
+    }
+
+    const isErc725Yv3 = await this.supportsInterface(
       address,
       ERC725Y_INTERFACE_IDS['3.0'],
     );
 
-    if (isErc725Y) {
-      return ERC725_VERSION.ERC725;
+    // The version 3 of the package can use the getData function from v2, still compatible
+    if (isErc725Yv3) {
+      return ERC725_VERSION.ERC725_v2;
     }
 
-    const isErc725Yv200 = await this.supportsInterface(
+    const isErc725Yv2 = await this.supportsInterface(
       address,
       ERC725Y_INTERFACE_IDS['2.0'],
     );
 
-    if (isErc725Yv200) {
-      return ERC725_VERSION.ERC725;
+    if (isErc725Yv2) {
+      return ERC725_VERSION.ERC725_v2;
     }
 
     // v0.2.0 and v0.6.0 have the same function signatures for getData, only versions before v0.2.0 requires a different call
@@ -197,8 +207,14 @@ export class ProviderWrapper {
     }
 
     switch (erc725Version) {
-      case ERC725_VERSION.ERC725:
-        return this._getAllData(address, keyHashes);
+      case ERC725_VERSION.ERC725_v5:
+        return this._getAllDataGeneric(
+          address,
+          keyHashes,
+          Method.GET_DATA_BATCH,
+        );
+      case ERC725_VERSION.ERC725_v2:
+        return this._getAllDataGeneric(address, keyHashes, Method.GET_DATA);
       case ERC725_VERSION.ERC725_LEGACY:
         return this._getAllDataLegacy(address, keyHashes);
       default:
@@ -206,20 +222,21 @@ export class ProviderWrapper {
     }
   }
 
-  private async _getAllData(
+  private async _getAllDataGeneric(
     address: string,
     keyHashes: string[],
+    method: Method.GET_DATA | Method.GET_DATA_BATCH,
   ): Promise<GetDataReturn[]> {
     if (this.type === ProviderTypes.ETHEREUM) {
       const encodedResults = await this.callContract(
         constructJSONRPC(
           address,
-          Method.GET_DATA,
+          method,
           abiCoder.encodeParameter('bytes32[]', keyHashes),
         ),
       );
 
-      const decodedValues = decodeResult(Method.GET_DATA, encodedResults);
+      const decodedValues = decodeResult(method, encodedResults);
 
       return keyHashes.map<GetDataReturn>((keyHash, index) => ({
         key: keyHash,
@@ -230,13 +247,13 @@ export class ProviderWrapper {
     const payload: JsonRpc[] = [
       constructJSONRPC(
         address,
-        Method.GET_DATA,
+        method,
         abiCoder.encodeParameter('bytes32[]', keyHashes),
       ),
     ];
 
     const results: any = await this.callContract(payload);
-    const decodedValues = decodeResult(Method.GET_DATA, results[0].result);
+    const decodedValues = decodeResult(method, results[0].result);
 
     return keyHashes.map<GetDataReturn>((key, index) => ({
       key,